Added Support for adding and getting from firestore database
This update allows for sample data to be inputted into the firestore database. The update also allows for the getting of the posted data from firestore and data changed on firestore changes on the web app as well. I will post an update soon on the structure of the database.